refactor: atomic_io infrastructure/ → utils/ 이동
proxy-cleanup Step B: atomic_io.py를 core/utils/atomic_io.py로 이동.
9개 소비자 + test_atomic_io.py 패치 경로 갱신.
import 정렬 자동 수정(I001). 행동 변경 없음.
