+++ core/cli/transcript.py
@@ -267,7 +267,7 @@ class SessionTranscript:
 
         import contextlib
 
-        from core.infrastructure.atomic_io import atomic_write_json
+        from core.utils.atomic_io import atomic_write_json
 
         with contextlib.suppress(OSError):
             atomic_write_json(index_path, index, indent=2)
