+++ core/utils/atomic_io.py
@@ -5,7 +5,7 @@ Pattern extracted from ``core/memory/session.py:_persist()``.
 
 Usage::
 
-    from core.infrastructure.atomic_io import atomic_write_text, atomic_write_json
+    from core.utils.atomic_io import atomic_write_text, atomic_write_json
 
     atomic_write_text(path, content)
     atomic_write_json(path, data, indent=2)
